requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
astunparse==1.6.3
attrdict==2.0.1
audioread==3.0.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
cookiecutter==2.6.0
DAWG-Python==0.7.2
decorator==5.1.1
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.110.0
flatbuffers==24.3.7
flet==0.21.1
flet-core==0.21.1
flet-runtime==0.21.1
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.6
importlib_metadata==7.0.2
jax==0.4.13
Jinja2==3.1.3
joblib==1.3.2
keras==2.12.0
lazy_loader==0.3
Levenshtein==0.25.0
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
MouseInfo==0.1.3
msgpack==1.0.8
mss==9.0.1
num2words==0.5.13
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pvporcupine==3.0.2
pvrecorder==1.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.0
python-slugify==8.0.4
pyttsx3==2.7
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.6.2
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sklearn==0.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.1
srt==3.5.3
starlette==0.36.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.3.0
tqdm==4.66.2
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
vosk==0.3.45
watchdog==4.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
Wikipedia-API==0.6.0
wrapt==1.14.1
zipp==3.18.0
